requirements.txt
Jinja2
python-dotenv
requests
starlette
uvicorn
bcrypt
cryptography
PyNaCl
GitPython
coloredlogs
blinker
background
